+++ src/cryptography/hazmat/backends/openssl/backend.py
@@ -2035,7 +2035,15 @@ class Backend:
         cipher_name = aead._aead_cipher_name(cipher)
         if self._fips_enabled and cipher_name not in self._fips_aead:
             return False
-        return self._lib.EVP_get_cipherbyname(cipher_name) != self._ffi.NULL
+        # SIV isn't loaded through get_cipherbyname but instead a new fetch API
+        # only available in 3.0+. But if we know we're on 3.0+ then we know
+        # it's supported.
+        if cipher_name.endswith(b"-siv"):
+            return self._lib.CRYPTOGRAPHY_OPENSSL_300_OR_GREATER == 1
+        else:
+            return (
+                self._lib.EVP_get_cipherbyname(cipher_name) != self._ffi.NULL
+            )
 
     @contextlib.contextmanager
     def _zeroed_bytearray(self, length: int) -> typing.Iterator[bytearray]: